SharedPreferences implemented and working, if logged in you will stay logged in.
